+++ caper/templates/pages/edit_project.html
@@ -79,7 +79,6 @@
                 <div id='result'></div>
                 <div style="border-top: 1px solid #000; margin: 20px 0;"></div>
 
-                <br>
             {% else %}
                 {% if field.name != 'accept_license' and field.name != 'replace_project' %}
                   {{field | as_crispy_field }}
@@ -131,8 +130,91 @@
           </div>
         </div>
 
+        
+
         <progress id="js-progressbar" class="uk-progress" value="0" max="100" hidden></progress>
         <div style="border-top: 1px solid #000; margin: 20px 0;"></div>
+
+
+        <div id="accordion">
+          <div class="card">
+              <div class="card-header" id="addMetadataHeader">
+                  <h5 class="mb-0">
+                      <button type="button" class="btn btn-link" data-toggle="collapse" data-target="#addMetadataContent" aria-expanded="false" aria-controls="addMetadataContent">
+                          Add Metadata
+                      </button>
+                  </h5>
+              </div>
+              <div id="addMetadataContent" class="collapse" aria-labelledby="addMetadataHeader" data-parent="#accordion">
+                  <div class="card-body">
+                      <h1>Add Metadata</h1>
+                      <p>
+                          Please upload a file in one of the following formats:
+                          <ul>
+                              <li><strong>CSV:</strong> Comma-separated values file.</li>
+                              <li><strong>TSV:</strong> Tab-separated values file.</li>
+                              <li><strong>XLSX:</strong> Excel spreadsheet file.</li>
+                          </ul>
+                      </p>
+                      <p>
+                          Ensure that:
+                          <ul>
+                              <li><strong>The first column</strong> is <b>"sample_name"</b> (e.g., "Sample_001").</li>
+                              <li><strong>The second column</strong> is <b>"cancer_type"</b> (e.g., "Lung Cancer").</li>
+                              <li><strong>All other columns</strong> are metadata fields associated with the sample.</li>
+                          </ul>
+                      </p>
+      
+                      <table class="table table-bordered table-striped">
+                          <thead class="thead-dark">
+                              <tr>
+                                  <th>sample_name</th>
+                                  <th>cancer_type</th>
+                                  <th>sample_type</th>
+                                  <th>tissue_source_site</th>
+                                  <th>race</th>
+                                  <th>ethnicity</th>
+                                  <th>tumor_stage</th>
+                                  <th>... (additional columns)</th>
+                              </tr>
+                          </thead>
+                          <tbody>
+                              <tr>
+                                  <td>Sample_001</td>
+                                  <td>Lung Cancer</td>
+                                  <td>Tumor</td>
+                                  <td>Lung</td>
+                                  <td>Asian</td>
+                                  <td>Hispanic</td>
+                                  <td>Stage II</td>
+                                  <td>...</td>
+                              </tr>
+                              <tr>
+                                  <td>Sample_002</td>
+                                  <td>Breast Cancer</td>
+                                  <td>Normal</td>
+                                  <td>Breast</td>
+                                  <td>White</td>
+                                  <td>Non-Hispanic</td>
+                                  <td>Stage I</td>
+                                  <td>...</td>
+                              </tr>
+                          </tbody>
+                      </table>
+      
+                      <h4>Note: Supported file types are <b>.csv</b>, <b>.tsv</b>, and <b>.xlsx</b>. Ensure your file is properly formatted!</h4>
+                      <hr>
+                      <div style="margin-bottom: 5px;">
+                          <label for="metadataFileInput">Upload Metadata File:</label>
+                          <input id="metadataFileInput" type="file" name="metadataFile" accept=".csv, .xlsx, .tsv">
+                      </div>
+                  </div>
+              </div>
+          </div>
+      </div>
+
+
+      <div style="border-top: 1px solid #000; margin: 20px 0;"></div>
         {{ run.accept_license | as_crispy_field }}
         
         <input
@@ -197,7 +279,7 @@
               addFileRow(file);
           }
       });
-  
+
       updateFileInput(); // Update the file input with the current filesArray
     });
   
@@ -254,6 +336,14 @@
       filesArray.forEach(file => {
         fd.append('document', file);
       });
+
+      const metadataFileInput = document.getElementById('metadataFileInput');
+
+      if (metadataFileInput.files.length > 0) {
+        const metadataFile = metadataFileInput.files[0];
+        console.log('hello');
+        fd.append('metadataFile', metadataFile);
+      };
   
       const xhr = new XMLHttpRequest();
       xhr.open('POST', uploadForm.action, true);
